% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_bet_to_entity}
\alias{convert_bet_to_entity}
\title{Convert a Worldbet string to HTML entities}
\usage{
convert_bet_to_entity(xs)
}
\arguments{
\item{xs}{a character vector of WorldBet strings}
}
\value{
a character vector of Unicode IPA symbols
}
\description{
Convert a Worldbet string to HTML entities
}
\examples{
convert_bet_to_entity("TIn")
#> [1] "θɪn"
convert_bet_to_ipa(c("TIn", "SUg&_r"))
#> [1] "θɪn"  "ʃʊgɚ"
}

