% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.R
\name{escape_regex}
\alias{escape_regex}
\title{Add escape slashes to a regex control character}
\usage{
escape_regex(xs)
}
\description{
Add escape slashes to a regex control character
}
\keyword{internal}

