% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldbet_key.R
\docType{data}
\name{worldbet_key}
\alias{worldbet_key}
\title{WorldBet symbol key}
\format{A data frame with 41 rows and 3 variables:
\describe{
  \item{WorldBet}{ASCII representation of the symbol}
  \item{IPA}{IPA symbol, encoded in UTF-8}
  \item{Entity}{HTML entity of the symbol}
}}
\usage{
worldbet_key
}
\description{
A table containing different representations of IPA symbols.
}
\keyword{datasets}

